+++ src/main/java/ru/smartel/strike/dto/response/event/EventDetailDTO.java
@@ -10,9 +10,9 @@ public class EventDetailDTO extends PostDetailDTO {
 
     private double latitude;
     private double longitude;
-    private int conflictId;
-    private Integer eventStatusId;
-    private Integer eventTypeId;
+    private long conflictId;
+    private Long eventStatusId;
+    private Long eventTypeId;
     private ConflictDetailDTO conflict;
     private ExtendedLocalityDTO locality;
 
@@ -45,27 +45,27 @@ public class EventDetailDTO extends PostDetailDTO {
         this.longitude = longitude;
     }
 
-    public int getConflictId() {
+    public long getConflictId() {
         return conflictId;
     }
 
-    public void setConflictId(int conflictId) {
+    public void setConflictId(long conflictId) {
         this.conflictId = conflictId;
     }
 
-    public Integer getEventStatusId() {
+    public Long getEventStatusId() {
         return eventStatusId;
     }
 
-    public void setEventStatusId(Integer eventStatusId) {
+    public void setEventStatusId(Long eventStatusId) {
         this.eventStatusId = eventStatusId;
     }
 
-    public Integer getEventTypeId() {
+    public Long getEventTypeId() {
         return eventTypeId;
     }
 
-    public void setEventTypeId(Integer eventTypeId) {
+    public void setEventTypeId(Long eventTypeId) {
         this.eventTypeId = eventTypeId;
     }
