+++ src/main/java/ru/smartel/strike/entity/Video.java
@@ -4,7 +4,15 @@ import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 import ru.smartel.strike.entity.reference.VideoType;
 
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
 import java.time.LocalDateTime;
 
 @Entity
@@ -13,7 +21,7 @@ public class Video {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column
-    private int id;
+    private long id;
 
     @CreationTimestamp
     @Column(name = "created_at")
@@ -33,7 +41,7 @@ public class Video {
     @JoinColumn(name = "video_type_id", referencedColumnName = "id")
     private VideoType videoType;
 
-    public int getId() {
+    public long getId() {
         return id;
     }
