+++ src/main/java/ru/smartel/strike/repository/client_version/ClientVersionRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ClientVersionRepository extends JpaRepository<ClientVersion, Integer> {
+public interface ClientVersionRepository extends JpaRepository<ClientVersion, Long> {
     Optional<ClientVersion> getByVersionAndClientId(String version, String clientId);
-    List<ClientVersion> findAllByIdGreaterThanAndClientId(int id, String clientId);
+    List<ClientVersion> findAllByIdGreaterThanAndClientId(long id, String clientId);
 }
