+++ src/main/java/ru/smartel/strike/repository/etc/IndustryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.smartel.strike.entity.reference.Industry;
 
 @Repository
-public interface IndustryRepository extends JpaRepository<Industry, Integer> {
+public interface IndustryRepository extends JpaRepository<Industry, Long> {
 }
