+++ src/main/java/ru/smartel/strike/repository/etc/TagRepository.java
@@ -7,6 +7,6 @@ import ru.smartel.strike.entity.Tag;
 import java.util.Optional;
 
 @Repository
-public interface TagRepository extends CrudRepository<Tag, Integer> {
+public interface TagRepository extends CrudRepository<Tag, Long> {
     Optional<Tag> findFirstByName(String name);
 }
