+++ src/main/java/ru/smartel/strike/rules/NotAParentEvent.java
@@ -4,10 +4,10 @@ import ru.smartel.strike.repository.event.EventRepository;
 
 public class NotAParentEvent extends BusinessRule {
 
-    private int eventId;
+    private long eventId;
     private EventRepository repository;
 
-    public NotAParentEvent(int eventId, EventRepository repository) {
+    public NotAParentEvent(long eventId, EventRepository repository) {
         this.eventId = eventId;
         this.repository = repository;
     }
